// backend/src/interfaces/status.interface.ts
export enum Status {
    SUCCESS = 'SUCCESS',
    CREATED = 'CREATED',
    FAILED_VALIDATION = 'FAILED_VALIDATION',
    UN_AUTHORIZED = 'UN_AUTHORIZED', 
    ERROR = 'ERROR',
    PROCESSING = 'PROCESSING',
    NOT_FOUND = 'NOT_FOUND',
    PRECONDITION_FAILED = 'PRECONDITION_FAILED',
    SUCCESS_NO_CONTENT = 'SUCCESS_NO_CONTENT', 
    FORBIDDEN = 'FORBIDDEN',
    UNPROCESSABLE_ENTRY = 'UNPROCESSABLE_ENTRY'
}